+++ src/main/java/com/song/shop/repository/ProductRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.song.shop.entity.ProductEntity;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> 
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> 
 {
 	List<ProductEntity> findByRegisterId(String registerId);
 }
